gpu: move array-expression lowering into its own unit
Turning whole-array work into element loops -- the self-alias temporary,
the section assignment, the strided-section gather a device pointer
cannot express, the elementwise assignment to an array variable -- plus
the small builders those lowerings share for extents, indices and
temporaries, is the last large group with a purpose of its own.

The include lists of gpu_offload.cpp and the new unit are pruned to what
each names, now that the split has made them different lists.

Pure code motion: the definitions are unchanged, only their home is.
